added a test of PWVD
