+++ main.py
@@ -4,7 +4,7 @@ import numpy as np
 import h5py
 import note_decompose as nde
 import note_recompose as nre
-import noise_gate_sigma as ngs
+import noise_gate as ngate
 import note_utils as note
 import sys
 import math_fun
@@ -12,12 +12,12 @@ import math_fun
 def main():
 	
 	print("decomposing data")
-	nde_class = nde.decompose('spoken.wav')
+	nde_class = nde.decompose('test.wav')
 	nde_class.octaves = (2,15)
 	nde_class.decompose('ns~test')
 
 	print("decomposing bg")
-	ngs_class = ngs.noise_gate_sigma('bg_fan.wav', nde_class.octaves)
+	ng_class = ngate.noise_gate_sigma('bg_fan.wav', nde_class.octaves)
 	
 		
 	fp = h5py.File('ns~test.hdf5', 'r', libver='latest')
@@ -34,8 +34,8 @@ def main():
 		freq = note.note2freq(int(key.split('-')[0]), key.split('-')[1])
 		print("noise gating: %s \t %.2f Hz " % (key, freq))
 	
-		d = ngs_class.noisegate(d, key, 2, spread=1000)
-		
+		d = ngs_class.noise_gate_sigma(d, key, 2, spread=1000)
+		#d = ngs_class.noise_gate_PWVD(d, spread=1000)
 
 		fp_out.create_dataset(key, data=d, dtype=d.dtype)
